requirements.txt
telebot
pandas
PyPDF2==3.0.1
python_dateutil==2.8.2
openpyxl==3.1.2
jinja2
pdfkit
load-dotenv